refactoring: app class에서 private field를 사용하지 않는 함수는 utils 로 이동
